Restart systemd unit if it crashes
There is a race condition (apparently when we scan for
wifi networks and bluetooth devices at the same time, even
by different processes). As a temporary workaround, we'll
just restart the service if it crashes.

We can use mutual exclusion to prevent two providers from
accessing the devices at the same time but we need to figure
out how to lock the devices from other process (NetworkManager).