Address race-condition for wireless device
This commit addresses the situation where two providers
attempt to scan for wireless devices (the nmea provider
scans for bluetooth devices while the wpsapi provider scans
for wifi APs).

It does this by making a mutex available to all providers
that must be locked before attempting to use the wifi
adapter.

It does not address the situation where another process
(ie. NetworkManager) scans for wifi hotspots while we're
scanning.